requirements.txt
#conda create --name qenv python=3.10
#conda activate qenv
#conda install -c pytorch -c nvidia faiss-gpu=1.8.0
#pip install -r requirements.txt

torch==2.5.1
transformers==4.48.1
datasets==3.2.0
tensorboard==2.18.0
nip-config==0.11.1
